add check user to all paths

+++ src/shared/modules/offer/offer.controller.ts
@@ -16,7 +16,6 @@ import {
   CommentRdo,
   CommentService,
   CreateCommentDTO,
-  CreateCommentRequest,
 } from '../comment/index.js';
 import {
   HttpMethod,
@@ -30,6 +29,26 @@ import {
 import { fillDTO } from '../../helpers/common.js';
 import { RestSchema } from '../../libs/config/rest.schema.js';
 import { Config } from '../../libs/config/config.interface.js';
+
+function buildOfferUpdateDto(
+  body: UpdateOfferDTO,
+  files?: {
+    previewImage?: Express.Multer.File[];
+    propertyPhotos?: Express.Multer.File[];
+  }
+): Partial<UpdateOfferDTO> {
+  const dto: Partial<UpdateOfferDTO> = { ...body };
+
+  if (files?.previewImage?.length) {
+    dto.previewImage = files.previewImage[0].path;
+  }
+
+  if (files?.propertyPhotos?.length) {
+    dto.propertyPhotos = files.propertyPhotos.map((f) => f.path);
+  }
+
+  return dto;
+}
 
 // TODO: add pagination to offers
 @injectable()
@@ -164,14 +183,16 @@ export class OfferController extends BaseController {
   }
 
   public async create(req: CreateOfferRequest, res: Response): Promise<void> {
-    const { body } = req;
+    const { body, tokenPayload } = req;
     const files = req.files as { [fieldname: string]: Express.Multer.File[] };
 
     const offerData = {
       ...body,
+      userId: tokenPayload.id,
       propertyPhotos: files?.propertyPhotos?.map((f) => f.path) ?? [],
       previewImage: files?.previewImage?.[0]?.path,
     };
+
     const offer = await this.offerService.create(offerData);
     const responseData = fillDTO(OfferRdo, offer);
     this.created(res, responseData);
@@ -203,15 +224,24 @@ export class OfferController extends BaseController {
   ): Promise<void> {
     const comments = await this.offerService.findComments(params.offerId);
     const responseData = fillDTO(CommentRdo, comments);
-    console.log({ comments, responseData });
+
     this.ok(res, responseData);
   }
 
   public async addComment(
-    { body }: CreateCommentRequest,
+    {
+      body,
+      tokenPayload,
+      params,
+    }: Request<ParamOfferId, unknown, CreateCommentDTO>,
     res: Response
   ): Promise<void> {
-    const comment = await this.commentService.create(body);
+    const { id: userId } = tokenPayload;
+    const comment = await this.commentService.create({
+      dto: body,
+      userId,
+      offerId: params.offerId,
+    });
     const responseData = fillDTO(CommentRdo, comment);
     this.created(res, responseData);
   }
@@ -220,8 +250,11 @@ export class OfferController extends BaseController {
     req: Request<ParamOfferId, unknown, UpdateOfferDTO>,
     res: Response
   ): Promise<void> {
-    const { params, body } = req;
+    const { params, body, tokenPayload } = req;
+
     const { offerId } = params;
+    const { id: userId } = tokenPayload;
+
     const files = req.files as
       | {
           previewImage?: Express.Multer.File[];
@@ -229,20 +262,11 @@ export class OfferController extends BaseController {
         }
       | undefined;
 
-    const updateDto: Partial<UpdateOfferDTO> = {
-      ...body,
-    };
-
-    if (files?.previewImage?.length) {
-      updateDto.previewImage = files.previewImage[0].path;
-    }
-
-    if (files?.propertyPhotos?.length) {
-      updateDto.propertyPhotos = files.propertyPhotos.map((f) => f.path);
-    }
+    const updateDto = buildOfferUpdateDto(body, files);
 
     const offer = await this.offerService.updateById({
       offerId,
+      userId,
       dto: updateDto,
     });
     const responseData = fillDTO(OfferRdo, offer);
@@ -251,11 +275,12 @@ export class OfferController extends BaseController {
   }
 
   public async delete(
-    { params }: Request<ParamOfferId>,
+    { params, tokenPayload }: Request<ParamOfferId>,
     res: Response
   ): Promise<void> {
     const { offerId } = params;
-    const offer = await this.offerService.deleteById(offerId);
+    const { id: userId } = tokenPayload;
+    const offer = await this.offerService.deleteById({ offerId, userId });
 
     await this.commentService.deleteByOfferId(offerId);
     const responseData = fillDTO(OfferRdo, offer);
